[WISE-244] moved WiseWebServiceException into org.jboss.wise.core.exception
